requirements.txt
Flask==3.0.3
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
# LLM clients (install only what you need)
openai==1.51.0
anthropic==0.34.2
Pillow==10.4.0
